Organise classes into more packages duke.task and duke.util
