+++ src/main/java/duke/task/Deadlines.java
@@ -1,4 +1,6 @@
-package duke;
+package duke.task;
+
+import duke.util.DateAndTimeParser;
 
 import java.time.LocalDate;
