+++ src/main/java/duke/task/TaskList.java
@@ -1,4 +1,7 @@
-package duke;
+package duke.task;
+
+import duke.util.DateAndTimeParser;
+import duke.DukeException;
 
 import java.io.BufferedReader;
 import java.io.IOException;
