+++ src/main/java/duke/task/Todos.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.task;
 
 public class Todos extends Task {
